backend/requirements.txt
fastapi
websockets
uvicorn[standard]
tortoise-orm
asyncpg
python-dotenv
